requirements.txt
python-binance
Telethon~=1.23.0
pytest
pygame